Balken kunnen nu niet meer uit het scherm verdwijnen. Object gemaakt van player, en veel variables veranderd naar short en floats ipv ints en doubles.

+++ Game9/Game9/Class1.cs
@@ -10,10 +10,10 @@ namespace Game9
     public class Ball
     {
         public int Direction { get; set; }
-        public double Speed { get; set; }
+        public float Speed { get; set; }
         public Vector2 Position { get; set; }
 
-        public Ball(int direction, double speed, Vector2 position) {
+        public Ball(int direction, float speed, Vector2 position) {
             Direction = direction;
             Speed = speed;
             Position = position;
